Add missing "px" after width and height value
I finished adding the value to both .avatar.proportioned
and avatar.distorted chained class elements without "px"
so the output does not change. I then looked for the answer
on TOP discord and found that I forgot to put "px" after
all of the width and heigh values. So I added them and the
output now work as I and the course expected to!