Edit .para class element
Para is the class for both the big and small paragraps
so I added the styles apply for both small and big
paragraph element, by putting the same styles in the
the .para body declaration. It would reduce the lines
of code and make the code easier to fix when a bug
related to the styles applied for both of the paragraphs
show up.