Add div before .child selector declartion
I added div declartion before .child to create chained
declaration for this element.

The style applied to this element would still work fine
in this situation but if there were more elements with
child class, this element would not be the only element
to have the style applied but every other elements within
child class.

So doing so would add the specification that only the
element that is descendant of <div> element and has
within child class.